new cart added

+++ app/Http/Controllers/Api/CartController.php
@@ -27,8 +27,7 @@ class CartController extends Controller
         $product = Product::where('slug', $slug)->first();
 
         if (!$product) {
-            return $this->error('Deal not found!', [], 404);
-
+            return response()->json(['error' => 'Deal not found!'], 404);
         }
 
         $customer_id = Auth::guard('api')->check() ? Auth::guard('api')->id() : null;
@@ -44,40 +43,43 @@ class CartController extends Controller
             $existing_cart = Cart::where('customer_id', $customer_id)->first();
             if ($existing_cart) {
                 if ($existing_cart->cart_number !== $cartnumber) {
+                    $new_cart = Cart::where('cart_number', $cartnumber)->first();
+                    if ($new_cart) {
+                        foreach ($new_cart->items as $item) {
+                            $existing_cart_item = CartItem::where('cart_id', $existing_cart->id)->where('product_id', $item->product_id)->first();
 
-                    $new_cart = Cart::where('cart_number', $cartnumber)->whereNull('customer_id')->first();
-
-                    foreach ($new_cart->items as $item) {
-                        $existing_cart_item = CartItem::where('cart_id', $existing_cart->id)->where('product_id', $item->product_id)->first();
-
-                        if ($existing_cart_item) {
-                            // If the item exists in both carts, increase the quantity
-                            $existing_cart_item->quantity += $item->quantity;
-                            $existing_cart_item->save();
-                        } else {
-                            // Assign new cart items to the existing cart
-                            $item->cart_id = $existing_cart->id;
-                            $item->save();
+                            if ($existing_cart_item) {
+                                // If the item exists in both carts, increase the quantity
+                                $existing_cart_item->quantity += $item->quantity;
+                                $existing_cart_item->save();
+                            } else {
+                                // Assign new cart items to the existing cart
+                                $item->cart_id = $existing_cart->id;
+                                $item->save();
+                            }
                         }
+
+                         // Update cart totals
+                        $existing_cart->item_count += $new_cart->item_count;
+                        $existing_cart->quantity += $new_cart->quantity;
+                        $existing_cart->total += $new_cart->total;
+                        $existing_cart->discount += $new_cart->discount;
+                        $existing_cart->shipping += $new_cart->shipping;
+                        $existing_cart->packaging += $new_cart->packaging;
+                        $existing_cart->handling += $new_cart->handling;
+                        $existing_cart->taxes += $new_cart->taxes;
+                        $existing_cart->grand_total += $new_cart->grand_total;
+                        $existing_cart->shipping_weight += $new_cart->shipping_weight;
+
+                        $existing_cart->save();
+
+                        $new_cart->delete();
+
+                        $old_cart = Cart::where('customer_id', $customer_id)->first();
+                    }else{
+                        $cartnumber = Str::uuid();
+                        $old_cart = Cart::where('customer_id', $customer_id)->first();
                     }
-
-                    // Update cart totals
-                    $existing_cart->item_count += $new_cart->item_count;
-                    $existing_cart->quantity += $new_cart->quantity;
-                    $existing_cart->total += $new_cart->total;
-                    $existing_cart->discount += $new_cart->discount;
-                    $existing_cart->shipping += $new_cart->shipping;
-                    $existing_cart->packaging += $new_cart->packaging;
-                    $existing_cart->handling += $new_cart->handling;
-                    $existing_cart->taxes += $new_cart->taxes;
-                    $existing_cart->grand_total += $new_cart->grand_total;
-                    $existing_cart->shipping_weight += $new_cart->shipping_weight;
-
-                    $existing_cart->save();
-
-                    $new_cart->delete();
-
-                    $old_cart = Cart::where('customer_id', $customer_id)->first();
                 } else {
                     $old_cart = Cart::where('customer_id', $customer_id)->first();
                 }
@@ -99,7 +101,7 @@ class CartController extends Controller
         if ($old_cart) {
             $item_in_cart = CartItem::where('cart_id', $old_cart->id)->where('product_id', $product->id)->first();
             if ($item_in_cart) {
-                return $this->error('Deal already in cart!', [], 400);
+                return response()->json(['error' => 'Deal already in cart!'], 400);
             }
         }
 
@@ -121,7 +123,7 @@ class CartController extends Controller
         $cart->ip_address = $request->ip();
         $cart->item_count = $old_cart ? ($old_cart->item_count + 1) : 1;
         $cart->quantity = $old_cart ? ($old_cart->quantity + $qtt) : $qtt;
-        $cart->total = $old_cart ? ($old_cart->total + ($product->original_price * $qtt)) : ($product->original_price * $qtt);
+        $cart->total = $old_cart ? ($old_cart->total + ($product->original_price  $qtt)) : ($product->original_price  $qtt);
         $cart->discount = $old_cart ? ($old_cart->discount + $discount) : $discount;
         $cart->shipping = $old_cart ? ($old_cart->shipping + $request->shipping) : $request->shipping;
         $cart->packaging = $old_cart ? ($old_cart->packaging + $request->packaging) : $request->packaging;
